Refactoring regarding how the infomation is displayed in the Home page
Quite huge refactoring on how the list items in Home page are displayed. Previously one anchor node and beacon was a separate list item (which was not correct..) now all four anchor nodes and one beacon is shown in one list item. This makes the beacon coordinates calculation more easy. Also, the code was reorganized in other places...